Session03

+++ departments/index.php
@@ -4,7 +4,7 @@ require_once 'C:/xampp/htdocs/finalDemo/app/dbconfig.php';
 // UI
 require_once "../shared/head.php";
 require_once "../shared/navbar.php";
-
+auth(2, 3);
 if (isset($_GET['delete'])) {
   $id = $_GET['delete'];
   $deleteQuery = "DELETE FROM `departments` WHERE id = $id";
@@ -27,7 +27,9 @@ $select = mysqli_query($con, $selectQuery);
           <tr>
             <th>#</th>
             <th>Department</th>
-            <th>Action</th>
+            <?php if ($_SESSION['user']['role'] == 1 || $_SESSION['user']['role'] == 2): ?>
+              <th>Action</th>
+            <?php endif; ?>
           </tr>
         </thead>
         <tbody>
@@ -35,10 +37,15 @@ $select = mysqli_query($con, $selectQuery);
             <tr>
               <td><?= $index + 1 ?></td>
               <td><?= $department['department'] ?></td>
-              <td>
-                <a href="edit.php?edit=<?= $department['id'] ?>" class="btn btn-warning">Edit</a>
-                <a href="?delete=<?= $department['id'] ?>" class="btn btn-danger">Delete</a>
-              </td>
+
+              <?php if ($_SESSION['user']['role'] == 1 || $_SESSION['user']['role'] == 2): ?>
+                <td>
+                  <?php if ($_SESSION['user']['role'] == 1): ?>
+                    <a href="edit.php?edit=<?= $department['id'] ?>" class="btn btn-warning">Edit</a>
+                  <?php endif; ?>
+                  <a href="?delete=<?= $department['id'] ?>" class="btn btn-danger">Delete</a>
+                </td>
+              <?php endif; ?>
             </tr>
           <?php endforeach; ?>
         </tbody>
